requirements/requirements.txt
asgiref==3.3.4
astroid==2.4.1
Django==3.2.3
django-easy-validator==1.5.3
django-environ==0.4.5
djangorestframework==3.12.4
dodgy==0.2.1
flake8==3.9.2
flake8-polyfill==1.0.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8-naming==0.10.0
prospector==1.3.1
pycodestyle==2.6.0
pydocstyle==6.1.1
pyflakes==2.2.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.1.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pytz==2021.1
PyYAML==5.4.1
requirements-detector==0.7
setoptconf==0.2.0
six==1.16.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
